move method from util to queryset
