+++ geo_areas/models.py
@@ -1,7 +1,9 @@
 from django.db import models
 from django.db.models import CheckConstraint
 from django.db.models import Max
+from django.db.models import OuterRef
 from django.db.models import Q
+from django.db.models import Subquery
 from polymorphic.managers import PolymorphicManager
 
 from common.business_rules import UniqueIdentifyingFields
@@ -23,6 +25,19 @@ from quotas import business_rules as quotas_business_rules
 class GeographicalAreaQuerySet(TrackedModelQuerySet):
     def erga_omnes(self):
         return self.filter(area_code=AreaCode.GROUP, area_id=1011)
+
+    def with_current_description(qs):
+        """Returns a GeographicalArea queryset annotated with the latest result
+        of a GeographicalAreaDescription subquery's description value, linking
+        these two queries on version_group field."""
+        current_descriptions = (
+            GeographicalAreaDescription.objects.current()
+            .filter(described_geographicalarea__version_group=OuterRef("version_group"))
+            .order_by("transaction__order")
+        )
+        return qs.annotate(
+            description=Subquery(current_descriptions.values("description")[:1]),
+        )
 
 
 class GeographicalArea(TrackedModel, ValidityMixin, DescribedMixin):
