+++ geo_areas/tests/test_models.py
@@ -3,6 +3,7 @@ import pytest
 from common.models.transactions import Transaction
 from common.models.utils import set_current_transaction
 from common.tests import factories
+from geo_areas import models
 
 pytestmark = pytest.mark.django_db
 
@@ -129,3 +130,21 @@ def test_geo_area_update_types(
     check_update_validation,
 ):
     assert check_update_validation(factory)
+
+
+def test_with_current_description():
+    """Tests that, after updating a geo area description,
+    with_current_description returns a queryset with one geo area annotated with
+    only the latest description."""
+    description = factories.GeographicalAreaDescriptionFactory.create(
+        description="blarghhh",
+    )
+    current_description = description.new_version(
+        description.transaction.workbasket,
+        description="bleurgh",
+    )
+    set_current_transaction(current_description.transaction)
+    qs = models.GeographicalArea.objects.current().with_current_description()
+
+    assert qs.count() == 1
+    assert qs.first().description == "bleurgh"
